relocated App component to src folder because i am tired of disrespect i am coming
