--- src/components/app/index.js
@@ -1,3 +1,0 @@
-import App from './App';
-
-export { App };
