+++ src/components/interface/index.js
@@ -1,5 +1,6 @@
 import Footer from './Footer/Footer';
 import Header from './Header/Header';
 import FormBlock from './FormBlock/FormBlock';
+import AppLoader from './AppLoader/AppLoader';
 
-export { Header, FormBlock, Footer };
+export { Header, FormBlock, Footer, AppLoader };
