+++ src/components/ui/index.js
@@ -1,5 +1,6 @@
 import TitleBlock from './TitleBlock/TitleBlock';
 import Spinner from './Spinner/Spinner';
 import Button from './Button/DefaultButton';
+import Preloader from './Preloader/Preloader';
 
-export { TitleBlock, Spinner, Button };
+export { TitleBlock, Spinner, Button, Preloader };
