+++ src/main.jsx
@@ -1,5 +1,6 @@
 import ReactDOM from 'react-dom/client';
-import { App } from './components/app';
+import { AppLoader } from './components/interface';
 import './styles/index.css';
 
-ReactDOM.createRoot(document.getElementById('root')).render(<App />);
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(<AppLoader />);
